+++ src/rpc/registry/actors/mod.rs
@@ -6,4 +6,6 @@ pub(crate) mod datacap;
 pub(crate) mod evm;
 pub(crate) mod init;
 pub(crate) mod miner;
+pub(crate) mod power;
+pub(crate) mod reward;
 pub(crate) mod system;